Django/Airbnb/pyproject.toml
[tool.poetry]
name = "airbnb"
version = "0.1.0"
description = ""
authors = ["kikiru328 <kikiru328@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
django = ">=5.1.7,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
pillow = ">=11.1.0,<12.0.0"
pyjwt = ">=2.10.1,<3.0.0"
django-environ = "^0.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
